handling error message

+++ src/data/explorer.py
@@ -1,3 +1,5 @@
+from sqlite3 import IntegrityError
+from error import Duplicate, Missing
 from . import curs
 from model.explorer import Explorer
 
@@ -16,7 +18,11 @@ def get_one(name: str) -> Explorer:
     qry = "select * from explorer where name =:name"
     params = {"name" : name}
     curs.execute(qry, params)
-    return row_to_model(curs.fetchone())
+    row = curs.fetchone()
+    if row:
+        return row_to_model(row)
+    else:
+        raise Missing(f"Explorer {name} not found")
 
 def get_all() -> list[Explorer] :
     qry = "select * from explorer"
@@ -27,10 +33,16 @@ def create(explorer: Explorer) -> Explorer:
     qry = """insert into explorer (name, country, description)
             values (:name, :country, :description)"""
     params = model_to_dict(explorer)
-    _ = curs.execute(qry, params)
+    try:
+        curs.execute(qry, params)
+    except IntegrityError:
+        raise Duplicate(msg=
+                        f"Explorer {explorer.name} already exists")    
     return get_one(explorer.name)
 
 def modify(name: str, explorer: Explorer) -> Explorer:
+    if not (name and explorer):
+        return None
     qry = """
         update explorer
         set country = :country,
@@ -39,12 +51,19 @@ def modify(name: str, explorer: Explorer) -> Explorer:
         where name = :name_orig """
     params = model_to_dict(explorer)
     params["name_orig"] = explorer.name
-    _ = curs.execute(qry, params)
-    explorer2 = get_one(explorer.name)
-    return explorer2
-
-def delete(explorer: Explorer) -> bool:
+    curs.execute(qry, params)
+    if curs.rowcount == 1:
+        return get_one(explorer.name)
+    else:
+        raise Missing(msg="Explorer {name} not found")
+    
+def delete(name: str):
+    if not name:
+        return False
     qry = "delete from explorer where name = :name"
-    params = {"name" : explorer.name}
-    res = curs.execute(qry, params)
-    return bool(res)
+    params = {"name" : name}
+    curs.execute(qry, params)
+    if curs.rowcount != 1:
+        raise Missing(msg=f"Explorer {name} not found")
+    return True
+
